+++ sound-mind-studio/tests/test_main_window.h
@@ -71,4 +71,5 @@ private slots:
     void deleteLayerRefusesToDeleteTheBackgroundLayer();
     void reorderLayersAppliesAValidPermutation();
     void reorderLayersRejectsAnInvalidPermutation();
+    void changingARealRowsOpacitySliderDoesNotCrash();
 };